Fixes a compile error in NewtMouseInput
git-svn-id: https://jmonkeyengine.googlecode.com/svn/trunk@10450 75d07b2b-3a1a-0410-a2c5-0572b91ccdca

+++ engine/src/jogl/com/jme3/input/jogl/NewtMouseInput.java
@@ -207,7 +207,7 @@ public class NewtMouseInput  implements MouseInput, MouseListener {
 
     public void mouseWheelMoved(MouseEvent awtEvt) {
         //FIXME not sure this is the right way to handle this case
-        int dwheel = awtEvt.getWheelRotation();
+        float dwheel = awtEvt.getWheelRotation();
         wheelPos += dwheel * WHEEL_AMP;
         cursorMoved = true;
     }
